requirements.txt
attrs==21.4.0
autopep8==1.6.0
Brotli==1.0.9
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.0.12
chart-studio==1.1.0
click==8.1.3
cloudpickle==2.1.0
colorama==0.4.4
convertdate==2.4.0
cryptography==37.0.2
cycler==0.11.0
dash==2.5.0
dash-auth==1.4.1
dash-bootstrap-components==1.1.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2022.6.0
distributed==2022.6.0
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Compress==1.12
flask-mongoengine==1.0.0
Flask-SeaSurf==1.1.1
Flask-WTF==1.0.1
fonttools==4.33.3
fsspec==2022.5.0
gunicorn==20.1.0
HeapDict==1.0.1
hijri-converter==2.2.4
holidays==0.14.2
htmlmin==0.1.12
idna==3.3
ImageHash==4.2.1
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.3
korean-lunar-calendar==0.2.1
locket==1.0.0
MarkupSafe==2.1.1
matplotlib==3.5.2
missingno==0.5.1
mongoengine==0.24.1
msgpack==1.0.4
multimethod==1.8
networkx==2.8.3
numpy==1.22.4
packaging==21.3
pandas==1.4.1
pandas-profiling==3.2.0
partd==1.2.0
passlib==1.7.4
phik==0.12.2
Pillow==9.1.1
plotly==5.8.2
psutil==5.9.1
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
pydotplus==2.0.2
PyJWT==2.4.0
PyMeeus==0.5.11
pymongo==4.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.0
retrying==1.3.3
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
sortedcontainers==2.4.0
tangled-up-in-unicode==0.2.0
tblib==1.7.0
tenacity==8.0.1
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm==4.64.0
typing_extensions==4.2.0
ua-parser==0.10.0
urllib3==1.26.9
visions==0.7.4
Werkzeug==2.1.2
woodwork==0.16.4
WTForms==3.0.1
zict==2.2.0
zipp==3.8.0
